[refactor]: refactored components and restructered folders
