+++ src/styles/App.css
@@ -1,0 +1,7 @@
+.App {
+  text-align: center;
+}
+
+.App-intro {
+  font-size: large;
+}
